Introduced Component and implemented first request in component
