+++ src/client/app/konvolut/konvolut.component.html
@@ -1,3 +1,7 @@
+<get-konvolut-iri [konvolut_id]="konvolut_id">
+
+</get-konvolut-iri>
+
 <div class="rae-convolute-view">
   <div class="rt-container">
     <div id="rt-main" class="mb9-sa3">
